+++ src/Symfony/Component/Validator/Constraints/RangeValidator.php
@@ -70,7 +70,30 @@ class RangeValidator extends ConstraintValidator
             }
         }
 
-        if (null !== $max && $value > $max) {
+        $hasLowerLimit = null !== $min;
+        $hasUpperLimit = null !== $max;
+
+        if ($hasLowerLimit && $hasUpperLimit && ($value < $min || $value > $max)) {
+            $violationBuilder = $this->context->buildViolation($constraint->notInRangeMessage)
+                ->setParameter('{{ value }}', $this->formatValue($value, self::PRETTY_DATE))
+                ->setParameter('{{ min }}', $this->formatValue($min, self::PRETTY_DATE))
+                ->setParameter('{{ max }}', $this->formatValue($max, self::PRETTY_DATE))
+                ->setCode(Range::NOT_IN_RANGE_ERROR);
+
+            if (null !== $constraint->maxPropertyPath) {
+                $violationBuilder->setParameter('{{ max_limit_path }}', $constraint->maxPropertyPath);
+            }
+
+            if (null !== $constraint->minPropertyPath) {
+                $violationBuilder->setParameter('{{ min_limit_path }}', $constraint->minPropertyPath);
+            }
+
+            $violationBuilder->addViolation();
+
+            return;
+        }
+
+        if ($hasUpperLimit && $value > $max) {
             $violationBuilder = $this->context->buildViolation($constraint->maxMessage)
                 ->setParameter('{{ value }}', $this->formatValue($value, self::PRETTY_DATE))
                 ->setParameter('{{ limit }}', $this->formatValue($max, self::PRETTY_DATE))
@@ -89,7 +112,7 @@ class RangeValidator extends ConstraintValidator
             return;
         }
 
-        if (null !== $min && $value < $min) {
+        if ($hasLowerLimit && $value < $min) {
             $violationBuilder = $this->context->buildViolation($constraint->minMessage)
                 ->setParameter('{{ value }}', $this->formatValue($value, self::PRETTY_DATE))
                 ->setParameter('{{ limit }}', $this->formatValue($min, self::PRETTY_DATE))